[CONFIGURATION] * NEW: Support for installing control icons in native components.
[64 BITS]
* BUG: Fixed some uses of long datatype in the gb.qt.ext component, the 
  Splitter and the Container class of gb.qt.
* BUG: Fixed some declarations in the gb.net.curl component.
* NEW: Pointer[] is a new class that represents an array of Pointer values.
* BUG: All gb.qt, gb.gtk and gb.desktop methods that deal with X11 Window 
  identifiers now use Pointer and Pointer[] datatypes instead of Integer 
  and Integer[].
* BUG: The Embedder example was updated according to the change above.

[INTERPRETER]
* BUG: Fixed the static array management that was broken by the 64 bits
  changes.

[GB.DEBUG]
* BUG: Debug.Write now does nothing if Debug.Start has not been called.

[GB.GTK]
* BUG: Fixed the implementation of the Message class that was broken.
* BUG: Fixed a crash in the Watcher class.
* BUG: Fixed the Image.Flip() and Image.Mirror() methods.
* BUG: Fixed a potential crash in timer management.
* NEW: The GridView Select event was implemented.


git-svn-id: svn://localhost/gambas/trunk@1033 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.net.curl/src/CCurl.h
@@ -75,13 +75,13 @@ typedef  struct
 void CCURL_stop(void *_object);
 
 void CCURL_stream_init  (GB_STREAM *stream,int fd);
-int  CCURL_stream_read  (GB_STREAM *stream, char *buffer, long len);
-int  CCURL_stream_write (GB_STREAM *stream, char *buffer, long len);
+int  CCURL_stream_read  (GB_STREAM *stream, char *buffer, int len);
+int  CCURL_stream_write (GB_STREAM *stream, char *buffer, int len);
 int  CCURL_stream_eof   (GB_STREAM *stream);
-int  CCURL_stream_lof   (GB_STREAM *stream, long long *len);
+int  CCURL_stream_lof   (GB_STREAM *stream, int64_t *len);
 int  CCURL_stream_open  (GB_STREAM *stream, const char *path, int mode, void *data);
-int  CCURL_stream_seek  (GB_STREAM *stream, long long pos, int whence);
-int  CCURL_stream_tell  (GB_STREAM *stream, long long *pos);
+int  CCURL_stream_seek  (GB_STREAM *stream, int64_t pos, int whence);
+int  CCURL_stream_tell  (GB_STREAM *stream, int64_t *pos);
 int  CCURL_stream_flush (GB_STREAM *stream);
 int  CCURL_stream_close (GB_STREAM *stream);
 int  CCURL_stream_handle(GB_STREAM *stream);
